removed and reorganized classes
