+++ app/modules/HomepageModule.scala
@@ -1,0 +1,12 @@
+package modules
+
+import controllers.HomepageController
+
+trait HomepageModule {
+  def homepageController: HomepageController
+}
+
+trait DefaultHomepageModuleImpl extends HomepageModule {
+  lazy val homepageController = new HomepageController
+}
+
